refactor: 카카오 로그인 관련 클래스 명칭을 socialLogin으로 변경
