+++ data/src/main/kotlin/com/sooum/data/di/RepositoryModule.kt
@@ -1,16 +1,16 @@
 package com.sooum.data.di
 
-import com.sooum.data.datastore.TokenProvider
+import com.sooum.data.network.TokenProvider
 import com.sooum.data.repository.AuthRepositoryImpl
 import com.sooum.data.repository.FriendRepositoryImpl
-import com.sooum.data.repository.KakaoRepositoryImpl
+import com.sooum.data.repository.SocialLoginRepositoryImpl
 import com.sooum.data.repository.MeetDetailCommentRepositoryImpl
 import com.sooum.data.repository.MeetDetailPlaceRepositoryImpl
 import com.sooum.data.repository.MeetDetailRepositoryImpl
 import com.sooum.data.repository.TokenProviderImpl
 import com.sooum.domain.repository.AuthRepository
 import com.sooum.domain.repository.FriendRepository
-import com.sooum.domain.repository.KakaoRepository
+import com.sooum.domain.repository.SocialLoginRepository
 import com.sooum.domain.repository.MeetDetailCommentRepository
 import com.sooum.domain.repository.MeetDetailPlaceRepository
 import com.sooum.domain.repository.MeetDetailRepository
@@ -58,8 +58,8 @@ abstract class RepositoryModule {
     @Binds
     @Singleton
     abstract fun bindKakaoRepository(
-        impl: KakaoRepositoryImpl
-    ): KakaoRepository
+        impl: SocialLoginRepositoryImpl
+    ): SocialLoginRepository
 
     @Binds
     @Singleton
