+++ data/src/main/kotlin/com/sooum/data/network/TokenProvider.kt
@@ -1,0 +1,7 @@
+package com.sooum.data.network
+
+interface TokenProvider {
+    fun getAccessToken(): String?
+    fun getRefreshToken(): String?
+    fun saveAccessToken(token: String)
+}
